server_code/webapp/requirements.txt
gunicorn
Flask
Jinja2
Werkzeug
distribute
wsgiref
tensorflow
keras
pillow
h5py
Plim
flask_mako
futures